refactor: extractor 상위 패키지로 옮김
