+++ src/main/java/persistence/sql/extractor/ColumnExtractor.java
@@ -1,10 +1,10 @@
-package persistence.sql.ddl.extractor;
+package persistence.sql.extractor;
 
 import jakarta.persistence.*;
 import persistence.sql.ddl.KeyType;
 import persistence.sql.ddl.dialect.Dialect;
-import persistence.sql.ddl.extractor.exception.ColumExtractorCreateException;
-import persistence.sql.ddl.extractor.exception.GenerationTypeMissingException;
+import persistence.sql.extractor.exception.ColumExtractorCreateException;
+import persistence.sql.extractor.exception.GenerationTypeMissingException;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
