+++ src/main/java/persistence/sql/extractor/exception/ColumExtractorCreateException.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.extractor.exception;
+package persistence.sql.extractor.exception;
 
 public class ColumExtractorCreateException extends RuntimeException {
     public ColumExtractorCreateException(final String message) {
