+++ src/test/java/persistence/sql/extractor/TableExtractorTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.extractor;
+package persistence.sql.extractor;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
